refactor: Tag 부분 전체적인 리팩토링
