+++ dev-timeline-core/src/main/java/com/sky7th/devtimeline/core/domain/tag/domain/Tag.java
@@ -1,9 +1,8 @@
-package com.sky7th.devtimeline.core.domain.tag;
+package com.sky7th.devtimeline.core.domain.tag.domain;
 
 import com.sky7th.devtimeline.core.domain.common.BaseTimeEntity;
-import com.sky7th.devtimeline.core.domain.post.Post;
-import com.sky7th.devtimeline.core.domain.post.linkpost.LinkPost;
-import com.sky7th.devtimeline.core.domain.user.User;
+import com.sky7th.devtimeline.core.domain.post.domain.Post;
+import com.sky7th.devtimeline.core.domain.user.domain.User;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
