+++ dev-timeline-core/src/main/java/com/sky7th/devtimeline/core/domain/tag/domain/TagRepository.java
@@ -1,4 +1,4 @@
-package com.sky7th.devtimeline.core.domain.tag;
+package com.sky7th.devtimeline.core.domain.tag.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
